refactor(zai): move the image-capability tests into their own module
`zai/tests.rs` is grandfathered at 1,685 lines, already 185 over the ratchet,
and appending the new vision coverage to it pushed it to 1,751. The guard's
own advice for growth that is not irreducible is to put the new code in a
module instead of raising the ceiling, so that is what this does: the file
returns to exactly its recorded ceiling and `zai/vision_tests.rs` holds the
75 lines that answer one self-contained question.

`command_deck.rs` does take its +20 as a ceiling raise. That growth is the
three-arm `match` on `route_mid_turn`, which has to sit at the call site
inside the driver's `select!` arm — the decision itself already moved out to
`subsession.rs`, which is where the testable part belongs.
