+++ stella-model/src/zai.rs
@@ -1463,3 +1463,5 @@ async fn aggregate_zai_stream(
 
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod vision_tests;
